requirements.txt
python-dotenv~=1.0.1
fastapi~=0.115.5
uvicorn
pydantic~=2.10.3
numpy~=1.26.4
requests~=2.32.3
langchain-core~=0.3.21
langchain-text-splitters~=0.3.2
pypdf~=5.1.0
langchain-google-genai~=2.0.6
openai~=1.56.1
rank-bm25~=0.2.2
transformers~=4.46.3
psycopg-binary
psycopg
python-multipart
tiktoken~=0.8.0
pdfplumber~=0.11.4
nltk
langchain-openai~=0.2.11
langchain-community~=0.3.9
langchain-experimental~=0.3.3
langchain_postgres
portkey_ai
langchain~=0.3.9